requirements.txt
# requirements.txt - UPDATED
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-multipart==0.0.6